pyproject.toml
[tool.poetry]
name = "gpress"
version = "0.1.0"
description = ""
authors = ["hdknr <gmail@hdknr.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
mysqlclient = "^1.4.6"
phpserialize = "^1.3"
ipython = "^7.13.0"
graphene-django = "^2.9.1"
djangorestframework = "^3.11.0"
django-filter = "^2.2.0"
django-cors-headers = "^3.2.1"
djangorestframework-filters = {git = "https://github.com/philipn/django-rest-framework-filters.git"}
click = "^7.1.2"
gunicorn = "^20.0.4"
channels = "^2.4.0"
channels_redis = "^3.1.0"
django-channels-graphql-ws = "^0.7.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
